Feat: make the cards and the tasks list working with thogeter and add in array
