+++ src/actions/cardsActions.js
@@ -1,0 +1,9 @@
+import { CONSTANTS } from "../actions"
+
+export const addCard = (listID, text) => {
+    //return object 
+    return {
+        type: CONSTANTS.ADD_CARD,
+        payload: {text, listID},
+    };
+};
