+++ src/actions/index.js
@@ -1,0 +1,7 @@
+export * from "./listActions";
+export * from "./cardsActions";
+
+export const CONSTANTS = {
+    ADD_CARD: "ADD_CARD",
+    ADD_LIST: "ADD_LIST"
+}
