+++ src/actions/listActions.js
@@ -1,0 +1,9 @@
+import { CONSTANTS } from "../actions"
+
+export const addList = (title) => {
+    //return object 
+    return {
+        type: CONSTANTS.ADD_LIST,
+        payload: title,
+    };
+};
